Half integer modules + Merging Tres' PR

+++ SDL/Event.cuh
@@ -169,8 +169,8 @@ namespace SDL
 
     extern struct modules* modulesInGPU;
     extern struct modules* modulesInHost;
-    extern unsigned int nModules;
-    extern unsigned int nLowerModules;
+    extern uint16_t nModules;
+    extern uint16_t nLowerModules;
     void initModules(const char* moduleMetaDataFilePath="data/centroid.txt"); //read from file and init
     void cleanModules();
     void initModulesHost(); //read from file and init
